requirements.txt
astroid==2.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.4
pydub==0.23.0
pylint==2.2.2
python-dotenv==0.10.0
scipy==1.1.0
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
